+++ tests/parse_from_file_spec.rs
@@ -14,7 +14,14 @@ use beancount_parser::BeancountFile;
 fn can_parse_example_files(#[case] file_name: &str, #[case] expected_directive_count: usize) {
     let mut path: PathBuf = "./tests/samples".into();
     path.push(file_name);
-    let mut file = BeancountFile::<f64>::default();
-    beancount_parser::read_files_v2([path], |entry| file.extend(Some(entry))).unwrap();
-    assert_eq!(file.directives.len(), expected_directive_count);
+    {
+        let mut file = BeancountFile::<f64>::default();
+        beancount_parser::read_files_v2([path.clone()], |entry| file.extend(Some(entry))).unwrap();
+        assert_eq!(file.directives.len(), expected_directive_count);
+    }
+    {
+        let mut file = BeancountFile::<f64>::default();
+        file.extend(beancount_parser::read_files_to_vec([path]).unwrap());
+        assert_eq!(file.directives.len(), expected_directive_count);
+    }
 }